+++ CHANGELOG.md
@@ -1,12 +1,29 @@
+## Version 1.7.0
+
+- **Features**
+  - Add the ability to bypass the intro screen that is shown before ticket creation screen.
+  - Add the ability to hide the DESK360 logo.
+
+- **Bug Fixes**
+  - Fix popover presentation on iPad.
+  - Fix ticket cell title layout on iPad.
+  - Fix privacy policy agreement text color.
+
+- **Breaking Changes**
+  - Introduced `Desk360Properties` to eliminate giant parameter list in `start` method.
+
 ## Version 1.6.0
+
 - **Minor Changes**
   - Update Moya dependency
 
 ## Version 1.5.1
+
 - **Minor Changes**
   - Changed the order of the args in `start` method.
 
 ## Version 1.5.0
+
 - **Features**
   - Add the ability to provide the current logged-in user name and email as default arguments.